+++ spec/lib/phobos/actions/process_message_spec.rb
@@ -11,10 +11,6 @@ RSpec.describe Phobos::Actions::ProcessMessage do
   class TestHandler2 < Phobos::EchoHandler
     include Phobos::Handler
 
-    def before_consume(payload, _metadata)
-      payload
-    end
-
     def self.around_consume(payload, metadata)
       yield payload, metadata
     end
@@ -22,10 +18,6 @@ RSpec.describe Phobos::Actions::ProcessMessage do
 
   class TestHandler3 < Phobos::EchoHandler
     include Phobos::Handler
-
-    def before_consume(payload)
-      payload
-    end
   end
 
   class TestHandler4 < Phobos::EchoHandler
@@ -67,8 +59,7 @@ RSpec.describe Phobos::Actions::ProcessMessage do
     allow(subject).to receive(:sleep) # Prevent sleeping in tests
   end
 
-  it 'processes the message by calling around consume, before consume and consume of the handler' do
-    expect(Phobos).not_to receive(:deprecate)
+  it 'processes the message by calling around consume and consume of the handler' do
     expect_any_instance_of(TestHandler).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
     expect_any_instance_of(TestHandler).to receive(:consume).with(payload, subject.metadata).once.and_call_original
 
@@ -89,63 +80,6 @@ RSpec.describe Phobos::Actions::ProcessMessage do
       )
     end
 
-    it 'supports and prefers around_consume if defined as a class method' do
-      expect(TestHandler2).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
-      expect(Phobos).to receive(:deprecate).twice # before_consume and around_consume as class method
-      expect_any_instance_of(TestHandler2).to receive(:before_consume).with(payload, subject.metadata).once.and_call_original
-      expect_any_instance_of(TestHandler2).to receive(:consume).with(payload, subject.metadata).once.and_call_original
-
-      subject.execute
-    end
-  end
-
-  context '#around_consume with nil message' do
-    let(:payload) { nil }
-
-    it 'should not deprecate with nil payload' do
-      expect(Phobos).not_to receive(:deprecate)
-      expect_any_instance_of(TestHandler).to receive(:around_consume).with(nil, subject.metadata).once.and_call_original
-      expect_any_instance_of(TestHandler).to receive(:consume).with(nil, subject.metadata).once.and_call_original
-
-      subject.execute
-    end
-  end
-
-  context '#around_consume that does not yield arguments' do
-    let(:listener) do
-      Phobos::Listener.new(
-        handler: TestHandler4,
-        group_id: 'test-group',
-        topic: topic
-      )
-    end
-
-    it 'supports the method and logs a deprecation message' do
-      expect(Phobos).to receive(:deprecate).once
-      expect_any_instance_of(TestHandler4).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
-      expect_any_instance_of(TestHandler4).to receive(:consume).with(payload, subject.metadata).once.and_call_original
-
-      subject.execute
-    end
-  end
-
-  context '#before_consume defined with 1 argument' do
-    let(:listener) do
-      Phobos::Listener.new(
-        handler: TestHandler3,
-        group_id: 'test-group',
-        topic: topic
-      )
-    end
-
-    it 'supports the method and logs a deprecation message' do
-      expect(Phobos).to receive(:deprecate).once
-      expect_any_instance_of(TestHandler3).to receive(:around_consume).with(payload, subject.metadata).once.and_call_original
-      expect_any_instance_of(TestHandler3).to receive(:before_consume).with(payload).once.and_call_original
-      expect_any_instance_of(TestHandler3).to receive(:consume).with(payload, subject.metadata).once.and_call_original
-
-      subject.execute
-    end
   end
 
   context 'with encoding' do